requirements.txt
algorithmia
python-dotenv